requirements.txt
# Django Framework
Django==5.2
django-background-tasks==1.2.8

# Document Processing
PyPDF2==3.0.1
docx2txt==0.8
pandas==2.2.0
chardet==5.2.0
openpyxl==3.1.2

# RAG and AI Libraries
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-ollama==0.1.0
langchain-huggingface==0.0.3
chromadb==0.4.24

# Embeddings and Vector Store
sentence-transformers==2.6.1
huggingface-hub==0.21.4

# HTTP Requests
requests==2.31.0

# Timezone handling
pytz==2024.1

# Development and Debugging
logging==0.4.9.6

# Optional: For production deployment
# gunicorn==21.2.0
# psycopg2-binary==2.9.9  # For PostgreSQL
# whitenoise==6.6.0       # For static files in production
